@@ -20,9 +20,9 @@
 		"framer-motion": "^6.2.8",
 		"multer": "^1.4.5-lts.1",
 		"nodemailer": "^6.8.0",
-		"react": "^17.0.2",
+		"react": "^18.2.0",
 		"react-bootstrap": "^2.5.0",
-		"react-dom": "^17.0.2",
+		"react-dom": "^18.2.0",
 		"react-router": "^6.4.1",
 		"react-router-dom": "^6.4.1",
 		"react-scripts": "^5.0.0",
